Move tickersHistory sync-schema to separate file
